small fixes, refactorings
