+++ contrib/ice/Makefile
@@ -1,5 +1,5 @@
 GOPATH0 := $(firstword $(subst :, ,$(GOPATH)))
 all:
-	GOFLAGS=$(GOFLAGS) go build -o $(GOPATH0)/bin/ca_importer reorg/importer/importer.go
+	GOFLAGS=$(GOFLAGS) go build -o $(GOPATH0)/bin/ice ice.go
 
 debug:
