+++ contrib/ice/ca-sync-job.sh
@@ -4,9 +4,9 @@ set -eEuo pipefail
 
 log_file="${CA_MKTSDB_PATH}/ingest-${now}.log"
 
-ca_importer \
-  -data_dir "${CA_MKTSDB_PATH}"
-  -reorg_dir "${CA_REORG_PATH}"
+ice reorg import \
+  "${CA_MKTSDB_PATH}"
+  "${CA_REORG_PATH}"
   -reimport "${CA_REIMPORT}" > "${log_file}"
 
 echo "Triggering marketstore restart"
